CORE-463 - Renames package reference to masterdata.
This is done for a better understanding and to avoid confusion about what reference means.

Original pull request: #146.
