+++ backend/src/main/java/quarano/department/web/TrackedCaseDiaryEntrySummary.java
@@ -3,7 +3,7 @@ package quarano.department.web;
 import lombok.RequiredArgsConstructor;
 import quarano.core.web.MapperWrapper;
 import quarano.diary.DiaryEntry;
-import quarano.reference.SymptomDto;
+import quarano.masterdata.SymptomDto;
 import quarano.tracking.ContactPerson;
 
 import java.time.LocalDateTime;
