+++ backend/src/main/java/quarano/masterdata/Symptom.java
@@ -1,4 +1,4 @@
-package quarano.reference;
+package quarano.masterdata;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
