+++ backend/src/main/java/quarano/masterdata/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Package for master data of the application which are a basis for the data model, but not a direct component. The data
+ * model references these basic data. The data in this package are either fixed or change completely independent of the
+ * domain-oriented data model.
+ */
+@org.springframework.lang.NonNullApi
+package quarano.masterdata;
