--- backend/src/main/java/quarano/reference/package-info.java
@@ -1,2 +1,0 @@
-@org.springframework.lang.NonNullApi
-package quarano.reference;
