+++ backend/src/test/java/quarano/department/web/EnrollmentWebIntegrationTests.java
@@ -19,8 +19,8 @@ import quarano.department.TrackedCase;
 import quarano.department.TrackedCaseRepository;
 import quarano.department.web.TrackedCaseRepresentations.DeviatingZipCode;
 import quarano.department.web.TrackedCaseRepresentations.InstitutionDto;
-import quarano.reference.Symptom;
-import quarano.reference.SymptomRepository;
+import quarano.masterdata.Symptom;
+import quarano.masterdata.SymptomRepository;
 import quarano.tracking.TrackedPerson.TrackedPersonIdentifier;
 import quarano.tracking.TrackedPersonDataInitializer;
 import quarano.tracking.TrackedPersonRepository;
