+++ backend/src/test/java/quarano/masterdata/SymptomControllerWebIntegrationTest.java
@@ -1,4 +1,4 @@
-package quarano.reference;
+package quarano.masterdata;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
@@ -6,6 +6,7 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import lombok.RequiredArgsConstructor;
 import quarano.QuaranoWebIntegrationTest;
 import quarano.WithQuaranoUser;
+import quarano.masterdata.SymptomController;
 
 import java.util.Locale;
